organized pages folder and fixed build errors
