+++ .gitignore
@@ -5,3 +5,5 @@ node_modules
 .env
 
 .vscode
+
+dist
