+++ src/web/src/Pages/upload/Upload.cy.tsx
@@ -1,4 +1,3 @@
-import React from 'react'
 import Upload from './Upload'
 
 describe('<Upload />', () => {
